Refactoring writer sevices
