+++ cmd/cassandra-writer/main.go
@@ -5,7 +5,6 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -14,7 +13,6 @@ import (
 	"strings"
 	"syscall"
 
-	"github.com/BurntSushi/toml"
 	kitprometheus "github.com/go-kit/kit/metrics/prometheus"
 	"github.com/gocql/gocql"
 	"github.com/mainflux/mainflux"
@@ -114,63 +112,13 @@ func loadConfig() config {
 	chanCfgPath := mainflux.Env(envChanCfgPath, defChanCfgPath)
 	subtCfgPath := mainflux.Env(envSubtCfgPath, defSubtCfgPath)
 	return config{
-		natsURL:  mainflux.Env(envNatsURL, defNatsURL),
-		logLevel: mainflux.Env(envLogLevel, defLogLevel),
-		port:     mainflux.Env(envPort, defPort),
-		dbCfg:    dbCfg,
-		channels: loadChansConfig(chanCfgPath),
-		subtopics: loadSubtopicsConfig(subtCfgPath),
+		natsURL:   mainflux.Env(envNatsURL, defNatsURL),
+		logLevel:  mainflux.Env(envLogLevel, defLogLevel),
+		port:      mainflux.Env(envPort, defPort),
+		dbCfg:     dbCfg,
+		channels:  writers.LoadChansConfig(chanCfgPath),
+		subtopics: writers.LoadSubtopicsConfig(subtCfgPath),
 	}
-}
-
-type filter struct {
-	List []string `toml:"filter"`
-}
-
-type chanConfig struct {
-	Channels filter `toml:"channels"`
-}
-
-func loadChansConfig(chanConfigPath string) map[string]bool {
-	data, err := ioutil.ReadFile(chanConfigPath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	var chanCfg chanConfig
-	if err := toml.Unmarshal(data, &chanCfg); err != nil {
-		log.Fatal(err)
-	}
-
-	chans := map[string]bool{}
-	for _, ch := range chanCfg.Channels.List {
-		chans[ch] = true
-	}
-
-	return chans
-}
-
-type subtConfig struct {
-	Subtopics filter `toml:"subtopics"`
-}
-
-func loadSubtopicsConfig(chanConfigPath string) map[string]bool {
-	data, err := ioutil.ReadFile(chanConfigPath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	var subtopicCfg subtConfig
-	if err := toml.Unmarshal(data, &subtopicCfg); err != nil {
-		log.Fatal(err)
-	}
-
-	subtopics := map[string]bool{}
-	for _, ch := range subtopicCfg.Subtopics.List {
-		subtopics[ch] = true
-	}
-
-	return subtopics
 }
 
 func connectToNATS(url string, logger logger.Logger) *nats.Conn {
