+++ cmd/influxdb-writer/main.go
@@ -5,14 +5,12 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
 
-	"github.com/BurntSushi/toml"
 	kitprometheus "github.com/go-kit/kit/metrics/prometheus"
 	influxdata "github.com/influxdata/influxdb/client/v2"
 	"github.com/mainflux/mainflux"
@@ -122,8 +120,8 @@ func loadConfigs() (config, influxdata.HTTPConfig) {
 		dbPort:    mainflux.Env(envDBPort, defDBPort),
 		dbUser:    mainflux.Env(envDBUser, defDBUser),
 		dbPass:    mainflux.Env(envDBPass, defDBPass),
-		channels:  loadChansConfig(chanCfgPath),
-		subtopics: loadSubtopicsConfig(subtCfgPath),
+		channels:  writers.LoadChansConfig(chanCfgPath),
+		subtopics: writers.LoadSubtopicsConfig(subtCfgPath),
 	}
 
 	clientCfg := influxdata.HTTPConfig{
@@ -133,56 +131,6 @@ func loadConfigs() (config, influxdata.HTTPConfig) {
 	}
 
 	return cfg, clientCfg
-}
-
-type filter struct {
-	List []string `toml:"filter"`
-}
-
-type chanConfig struct {
-	Channels filter `toml:"channels"`
-}
-
-func loadChansConfig(chanConfigPath string) map[string]bool {
-	data, err := ioutil.ReadFile(chanConfigPath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	var chanCfg chanConfig
-	if err := toml.Unmarshal(data, &chanCfg); err != nil {
-		log.Fatal(err)
-	}
-
-	chans := map[string]bool{}
-	for _, ch := range chanCfg.Channels.List {
-		chans[ch] = true
-	}
-
-	return chans
-}
-
-type subtConfig struct {
-	Subtopics filter `toml:"subtopics"`
-}
-
-func loadSubtopicsConfig(subtopicConfigPath string) map[string]bool {
-	data, err := ioutil.ReadFile(subtopicConfigPath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	var subtopicCfg subtConfig
-	if err := toml.Unmarshal(data, &subtopicCfg); err != nil {
-		log.Fatal(err)
-	}
-
-	subtopics := map[string]bool{}
-	for _, ch := range subtopicCfg.Subtopics.List {
-		subtopics[ch] = true
-	}
-
-	return subtopics
 }
 
 func makeMetrics() (*kitprometheus.Counter, *kitprometheus.Summary) {
